requirements.txt
django>=4.2.7
djangorestframework>=3.14.0
mysqlclient>=2.2.0
django-cors-headers>=4.3.1
